+++ Runtime/AuthenticationModel.cs
@@ -282,6 +282,7 @@ namespace com.noctuagames.sdk
                 {
                     purchaseItemName = item.OrderRequest.ProductId;
                 }
+
                 _pendingPurchases.Add(new PendingPurchaseItem{
                     OrderId = item.OrderId,
                     Timestamp = item.OrderRequest.Timestamp,
@@ -290,7 +291,7 @@ namespace com.noctuagames.sdk
                     PurchaseItemName = purchaseItemName,
                     OrderRequest = Convert.ToBase64String(Encoding.UTF8.GetBytes(JsonConvert.SerializeObject(item.OrderRequest))),
                     VerifyOrderRequest = Convert.ToBase64String(Encoding.UTF8.GetBytes(JsonConvert.SerializeObject(item.VerifyOrderRequest))),
-                    PlayerId = GetPlayerIdFromJwt(item.AccessToken)
+                    PlayerId = item.PlayerId,
                 });
             }
 
@@ -307,58 +308,6 @@ namespace com.noctuagames.sdk
             return await _iapService.RetryPendingPurchaseByOrderId(orderId);
         }
         
-        private long? GetPlayerIdFromJwt(string token)
-        {
-            if (string.IsNullOrEmpty(token))
-            {
-                _log.Warning("Empty JWT token.");
-                
-                return null;
-            }
-
-            var parts = token.Split('.');
-            if (parts.Length != 3)
-            {
-                _log.Error("Invalid JWT format: Token must have 3 parts (header, payload, signature).");
-                
-                return null;
-            }
-            
-            try
-            {
-                var payload = parts[1];
-                var paddedPayload = payload.PadRight((payload.Length + 2) / 3 * 3, '=');                
-                var json = Encoding.UTF8.GetString(Convert.FromBase64String(paddedPayload));
-                var parsed = JObject.Parse(json);
-
-                if (parsed["player_id"] == null)
-                {
-                    _log.Error("Missing 'player_id' claim in JWT payload.");
-                    
-                    return null;
-                }
-
-                var playerId = parsed["player_id"]?.Value<long?>();
-                
-                if (playerId == null)
-                {
-                    _log.Error("The 'player_id' claim is not a valid number.");
-                }
-
-                return playerId;
-            }
-            catch (FormatException ex)
-            {
-                _log.Error($"Base64 decoding error: {ex.Message}");
-            }
-            catch (Exception ex)
-            {
-                _log.Error($"Unexpected error while parsing JWT: {ex.Message}");
-            }
-
-            return null;
-        }
-
     }
     
     internal enum AuthIntention